+++ lib/id_pair_set.dart
@@ -1,5 +1,7 @@
 /// A library for managing sets of identifier pairs.
 library;
 
+export 'src/exceptions.dart';
 export 'src/id_pair.dart';
 export 'src/id_pair_set.dart';
+export 'src/id_registry.dart';